Add a global player (header player)
Includes the ability for metadata to be stored with a track.
The type of such metadata is specified with a generic, as to keep it out of the TrackManager component.

This is slightly awkward because the TrackManager singleton must of of a single type, but that type is not known.